Update to v88.2
App is now out of date

+++ common/post-fs-data.sh
@@ -1,7 +1,7 @@
 DISABLE=/data/data/de.robv.android.xposed.installer/conf/disabled
 MIRRDIR=/dev/magisk/mirror
 
-[[ -f $DISABLE && $SH != "/magisk/"* ]] && exit
+[[ -f $DISABLE || $SH != "/magisk/"* ]] && exit
 
 mount -o rw,remount /
 ln -s $SH/xposed.prop /xposed.prop
